Replacing call to pi_gregory_pauseable_2
There were some issues with the storing and passing of pointers between
functions. At some point the pointers were invalidated.

It makes more sense to have the calculation in the main, that way is
easier to debug/fix/understand the code and reduce the complication
of variables manipulation.

With this approach, segmentation faults were reduced to 0. The only
downside of the actual implementation is that we can't achieve a very
exact Pi approximation.